+++ src/main/kotlin/feature/budget/domain/model/BudgetStatus.kt
@@ -1,9 +1,11 @@
 package feature.budget.domain.model
 
+import java.math.BigDecimal
+
 data class BudgetStatus(
-    val limit: Double,
-    val spent: Double,
-    val remaining: Double,
+    val limit: BigDecimal,
+    val spent: BigDecimal,
+    val remaining: BigDecimal,
     val percentageUsed: Double,
     val isExceeded: Boolean
 )
